chore(crdt-historical): move table header to new file
