+++ src/components/pages/state/race-ethnicity/historical-tables.module.scss
@@ -53,20 +53,6 @@
   }
 }
 
-.table-header {
-  color: $color-plum-700;
-  background-color: $color-plum-100;
-
-  @include margin(0, (bottom, top));
-  @include padding(4);
-
-  text-align: left;
-}
-
-th.hidden * {
-  display: none;
-}
-
 .abbreviated-label {
   color: $color-plum-700;
   text-decoration: underline;
